Add format-specifier warnings to SkDebugf.
This CL fixes up many existing format-specifier violations in Skia.
Note that GCC has a warning for formatting nothing, so existing calls to
`SkDebugf("")` have been removed, or replaced with `SkDebugf("%s", "")`.
These were apparently meant to be used as a place to set a breakpoint.

Some of our clients also use SkDebug with bad format specifiers, so this
check is currently only enabled when SKIA_IMPLEMENTATION is true.

+++ experimental/skrive/src/Artboard.cpp
@@ -48,7 +48,7 @@ std::tuple<sk_sp<Component>, size_t> parse_component(StreamReader* sr) {
             break;
     }
 
-    SkDebugf("!! unsupported node type: %d\n", block.type());
+    SkDebugf("!! unsupported node type: %d\n", (int)block.type());
     return {nullptr, 0};
 }
 
@@ -109,7 +109,7 @@ sk_sp<Artboard> parse_artboard(StreamReader* sr) {
             ab->setRoot(parse_components(sr));
             break;
         default:
-            SkDebugf("!! Unsupported block type: %d\n", block.type());
+            SkDebugf("!! Unsupported block type: %d\n", (int)block.type());
             break;
         }
     }
