+++ src/gpu/vk/GrVkImage.h
@@ -17,6 +17,8 @@
 #include "src/gpu/GrManagedResource.h"
 #include "src/gpu/GrTexture.h"
 
+#include <cinttypes>
+
 class GrVkGpu;
 class GrVkTexture;
 
@@ -193,7 +195,7 @@ private:
 
 #ifdef SK_TRACE_MANAGED_RESOURCES
         void dumpInfo() const override {
-            SkDebugf("GrVkImage: %d (%d refs)\n", fImage, this->getRefCnt());
+            SkDebugf("GrVkImage: %" PRIdPTR " (%d refs)\n", (intptr_t)fImage, this->getRefCnt());
         }
 #endif
